requirements.txt
fastapi
uvicorn[standard]
pyjwt
python-dotenv
psycopg[binary]
psycopg-pool

